[#289]add junit test for pr
