+++ sentinel-adapter/sentinel-grpc-adapter/src/test/java/com/alibaba/csp/sentinel/adapter/grpc/FooServiceImpl.java
@@ -27,10 +27,7 @@ class FooServiceImpl extends FooServiceGrpc.FooServiceImplBase {
 
     @Override
     public void sayHello(FooRequest request, StreamObserver<FooResponse> responseObserver) {
-        if (request.getId() == -1) {
-            responseObserver.onError(new IllegalAccessException("The id is error"));
-            return;
-        }
+
         String message = String.format("Hello %s! Your ID is %d.", request.getName(), request.getId());
 
         FooResponse response = FooResponse.newBuilder().setMessage(message).build();
@@ -40,6 +37,25 @@ class FooServiceImpl extends FooServiceGrpc.FooServiceImplBase {
 
     @Override
     public void anotherHello(FooRequest request, StreamObserver<FooResponse> responseObserver) {
+
+        String message = String.format("Good day, %s (%d)", request.getName(), request.getId());
+        FooResponse response = FooResponse.newBuilder().setMessage(message).build();
+        responseObserver.onNext(response);
+        responseObserver.onCompleted();
+    }
+    @Override
+    public void helloWithEx(FooRequest request, StreamObserver<FooResponse> responseObserver) {
+        if (request.getId() == -1) {
+            responseObserver.onError(new IllegalAccessException("The id is error"));
+            return;
+        }
+        String message = String.format("Good day, %s (%d)", request.getName(), request.getId());
+        FooResponse response = FooResponse.newBuilder().setMessage(message).build();
+        responseObserver.onNext(response);
+        responseObserver.onCompleted();
+    }
+    @Override
+    public void anotherHelloWithEx(FooRequest request, StreamObserver<FooResponse> responseObserver) {
         if (request.getId() == -1) {
             responseObserver.onError(new IllegalAccessException("The id is error"));
             return;
