+++ sentinel-adapter/sentinel-grpc-adapter/src/test/java/com/alibaba/csp/sentinel/adapter/grpc/SentinelGrpcClientInterceptorTest.java
@@ -15,22 +15,20 @@
  */
 package com.alibaba.csp.sentinel.adapter.grpc;
 
+import java.util.Collections;
+
 import com.alibaba.csp.sentinel.EntryType;
 import com.alibaba.csp.sentinel.adapter.grpc.gen.FooRequest;
 import com.alibaba.csp.sentinel.adapter.grpc.gen.FooResponse;
 import com.alibaba.csp.sentinel.node.ClusterNode;
 import com.alibaba.csp.sentinel.slots.block.RuleConstant;
-import com.alibaba.csp.sentinel.slots.block.degrade.DegradeRule;
-import com.alibaba.csp.sentinel.slots.block.degrade.DegradeRuleManager;
 import com.alibaba.csp.sentinel.slots.block.flow.FlowRule;
 import com.alibaba.csp.sentinel.slots.block.flow.FlowRuleManager;
 import com.alibaba.csp.sentinel.slots.clusterbuilder.ClusterBuilderSlot;
+
 import io.grpc.StatusRuntimeException;
 import org.junit.After;
 import org.junit.Test;
-
-import java.io.IOException;
-import java.util.Collections;
 
 import static org.junit.Assert.*;
 
@@ -44,32 +42,16 @@ public class SentinelGrpcClientInterceptorTest {
     private final String resourceName = "com.alibaba.sentinel.examples.FooService/sayHello";
     private final int threshold = 2;
     private final GrpcTestServer server = new GrpcTestServer();
-    private final int timeWindow = 10;
 
     private void configureFlowRule(int count) {
         FlowRule rule = new FlowRule()
-                .setCount(count)
-                .setGrade(RuleConstant.FLOW_GRADE_QPS)
-                .setResource(resourceName)
-                .setLimitApp("default")
-                .as(FlowRule.class);
+            .setCount(count)
+            .setGrade(RuleConstant.FLOW_GRADE_QPS)
+            .setResource(resourceName)
+            .setLimitApp("default")
+            .as(FlowRule.class);
         FlowRuleManager.loadRules(Collections.singletonList(rule));
     }
-
-
-    private void configureDegradeRule(int count) {
-        DegradeRule rule = new DegradeRule()
-                .setCount(count)
-                .setGrade(RuleConstant.DEGRADE_GRADE_EXCEPTION_COUNT)
-                .setResource(resourceName)
-                .setLimitApp("default")
-                .as(DegradeRule.class)
-                .setTimeWindow(timeWindow);
-        DegradeRuleManager.loadRules(Collections.singletonList(rule));
-
-
-    }
-
 
     @Test
     public void testGrpcClientInterceptor() throws Exception {
@@ -98,38 +80,6 @@ public class SentinelGrpcClientInterceptorTest {
     private boolean sendRequest(FooServiceClient client) {
         try {
             FooResponse response = client.sayHello(FooRequest.newBuilder().setName("Sentinel").setId(666).build());
-            System.out.println("Response: " + response);
-            return true;
-        } catch (StatusRuntimeException ex) {
-            System.out.println("Blocked, cause: " + ex.getMessage());
-            return false;
-        }
-    }
-
-
-    @Test
-    public void testGrpcClientInterceptor_degrade() throws IOException {
-        final int port = 19328;
-
-        configureDegradeRule(1);
-        server.start(port, false);
-
-        FooServiceClient client = new FooServiceClient("localhost", port, new SentinelGrpcClientInterceptor());
-
-        assertFalse(sendErrorRequest(client));
-        ClusterNode clusterNode = ClusterBuilderSlot.getClusterNode(resourceName, EntryType.OUT);
-        assertNotNull(clusterNode);
-        assertEquals(1, clusterNode.exceptionQps());
-        // The second request will be blocked.
-        assertFalse(sendRequest(client));
-        assertEquals(1, clusterNode.blockRequest());
-
-        server.stop();
-    }
-
-    private boolean sendErrorRequest(FooServiceClient client) {
-        try {
-            FooResponse response = client.sayHello(FooRequest.newBuilder().setName("Sentinel").setId(-1).build());
             System.out.println("Response: " + response);
             return true;
         } catch (StatusRuntimeException ex) {
