+++ sentinel-adapter/sentinel-grpc-adapter/src/test/proto/example.proto
@@ -17,7 +17,14 @@ message FooResponse {
 
 // Example service definition.
 service FooService {
+
+
     rpc sayHello(FooRequest) returns (FooResponse) {}
 
     rpc anotherHello(FooRequest) returns (FooResponse) {}
+
+    rpc helloWithEx(FooRequest) returns (FooResponse) {}
+    rpc anotherHelloWithEx(FooRequest) returns (FooResponse) {}
+
+
 }
